src/main/java/leetcode/easy206/Solution.java
package leetcode.easy206;

/**
 * 206. Reverse Linked List
 * https://leetcode.com/problems/reverse-linked-list/description/
 */
class Solution {
    public ListNode reverseList(ListNode head) {

        ListNode prev = null;
        while (head != null) {
            ListNode tmp = head.next;
            head.next = prev;
            prev = head;
            if (tmp == null) {
                return head;
            }
            head = tmp;
        }
        return head;
    }

    public static class ListNode {
        int val;
        ListNode next;

        ListNode(int x) {
            val = x;
        }
    }
}
